backend/requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
paramiko>=3.0.0
SimpleITK>=2.3.0
nibabel>=5.0.0
pydicom>=2.4.4,<3.0.0
numpy<2.0
vtk>=9.2.0
matplotlib>=3.7.0
scipy>=1.10.0
